\name{kldata}
\alias{kldata}
\non_function{}
\title{The data set given in Kalbfleisch and Lawless (1985). }
\usage{data(kldata)}

}
\source{
  Kalbfleisch, J.D. and Lawless, J.F. `The analysis
of panel data under a Markov assumption', {\em Journal of the
American Statistical Association}, {\bf 80}, 863-871, (1985).
}
\description{
In Section 6 of Kalbfleisch and Lawless (1985) some data simulated from
a three-state Markov model is given.
They started 15 individuals in each of States 1 and 2 for each of four
groups.Individuals were followed for 5 transitions at equidistant time
intervals. 
  }
\examples{
data(kldata)
##see the user's manual for explicit examples
}
\keyword{datasets}
